added new enemy type + organized players into parent object
added new enemy type:
shooting enemy. it stops when it sees a player and shoots at them. currently the shooting is NOT functional. will add that later

moved all shared player code into "oParentCharacter" for ease of future programming. i was running into issues where i had to code enemies twice to interact with the pc so this should simplify that